requirements.txt
# ============================================
# STREAMLIT & CORE DEPENDENCIES
# ============================================
streamlit==1.37.0
pandas==2.2.2
numpy==1.26.2
openpyxl==3.1.5
filelock==3.15.4
PyYAML==6.0.2

# ============================================
# KAFKA & MESSAGING
# ============================================
kafka-python==2.0.2

# ============================================
# DATABASE
# ============================================
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.23

# ============================================
# MONITORING
# ============================================
prometheus-client==0.18.0

# ============================================
# CONFIGURATION & UTILITIES
# ============================================
python-dotenv==1.0.0
requests==2.31.0

# ============================================
# VISUALIZATION
# ============================================
matplotlib==3.8.0
plotly==5.18.0